requirements.txt
pillow
pyside6
